requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain>=0.1.0
langchain-groq>=0.1.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
pydantic==2.5.2
python-dotenv==1.0.0
python-multipart==0.0.6

